Simplify bin script, introduce Initializer
